Add very basic element visitor implementations

+++ modules/code-gen/src/main/java/net/strokkur/jap/code/expression/Expressions.java
@@ -60,7 +60,7 @@ public final class Expressions {
     return new CodeIntExpression(value);
   }
 
-  public static CodeLongExpression longExpr(int value) {
+  public static CodeLongExpression longExpr(long value) {
     return new CodeLongExpression(value);
   }
